requirements.txt
# --- Core Framework ---
Django==5.2.7
djangorestframework==3.16.1

# --- Database ---
dj-database-url==3.0.1
psycopg2-binary==2.9.11

# --- Web Server & Static Files ---
gunicorn==23.0.0
whitenoise==6.11.0

# --- Cloud Storage (AWS S3) ---
boto3==1.40.59
django-storages==1.14.6

# --- Configuration ---
python-dotenv==1.1.1

# --- Image Processing ---
pillow==11.3.0